requirements.txt
streamlit
opencv-python
numpy
Pillow
cvzone